Rename signing_tests to ciphersuite_generic_tests to match frost-core (#364)
