+++ frost-ed25519/src/tests.rs
@@ -1,12 +1,12 @@
 #[cfg(test)]
 mod batch_tests;
 #[cfg(test)]
+mod ciphersuite_generic_tests;
+#[cfg(test)]
 mod coefficient_commitment_tests;
 #[cfg(test)]
 mod deserialize_tests;
 #[cfg(test)]
 mod proptests;
-#[cfg(test)]
-mod signing_tests;
 #[cfg(test)]
 mod vss_commitment_tests;
